[ refactor ] Move conversion from clauses to rewrite rules to own module
